chore: fix broken tests and imports
