+++ packages/appkit/exports/react.ts
@@ -14,7 +14,7 @@ export * from '../src/library/react/index.js'
 export * from '../src/utils/index.js'
 export type * from '@reown/appkit-core'
 export type { CaipNetwork, CaipAddress, CaipNetworkId } from '@reown/appkit-common'
-export { CoreHelperUtil, AccountController, NetworkController } from '@reown/appkit-core'
+export { CoreHelperUtil, AccountController } from '@reown/appkit-core'
 
 export let modal: AppKit | undefined = undefined
