+++ packages/appkit/src/universal-adapter/client.ts
@@ -4,7 +4,6 @@ import {
   ChainController,
   ConnectionController,
   CoreHelperUtil,
-  NetworkController,
   StorageUtil,
   type ConnectionControllerClient,
   type Connector,
@@ -397,15 +396,13 @@ export class UniversalAdapterClient {
 
       const storedCaipNetwork = StorageUtil.getStoredActiveCaipNetwork()
       const activeCaipNetwork = ChainController.state.activeCaipNetwork
-
+      const approvedCaipNetworkIds = ChainController.getApprovedCaipNetworkIds()
       try {
         if (storedCaipNetwork) {
           ChainController.setActiveCaipNetwork(storedCaipNetwork)
         } else if (!activeCaipNetwork) {
           this.setDefaultNetwork(nameSpaces)
-        } else if (
-          !NetworkController.state.approvedCaipNetworkIds?.includes(activeCaipNetwork.id)
-        ) {
+        } else if (!approvedCaipNetworkIds?.includes(activeCaipNetwork.id)) {
           this.setDefaultNetwork(nameSpaces)
         }
       } catch (error) {
@@ -427,7 +424,7 @@ export class UniversalAdapterClient {
         const chainId = namespace.chains[0]
 
         if (chainId) {
-          const requestedCaipNetworks = NetworkController.state?.requestedCaipNetworks
+          const requestedCaipNetworks = ChainController.getRequestedCaipNetworks()
 
           if (requestedCaipNetworks) {
             const network = requestedCaipNetworks.find(c => c.id === chainId)
