+++ packages/core/src/utils/TypeUtil.ts
@@ -10,10 +10,6 @@ import type {
   SdkFramework,
   AppKitSdkVersion
 } from '@reown/appkit-common'
-import type {
-  NetworkControllerClient,
-  NetworkControllerState
-} from '../controllers/NetworkController.js'
 import type { ConnectionControllerClient } from '../controllers/ConnectionController.js'
 import type { AccountControllerState } from '../controllers/AccountController.js'
 import type { OnRampProviderOption } from '../controllers/OnRampController.js'
@@ -823,6 +819,23 @@ export interface WriteContractArgs {
   method: 'send' | 'transfer' | 'call'
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   abi: any
+}
+
+export interface NetworkControllerClient {
+  switchCaipNetwork: (network: CaipNetwork) => Promise<void>
+  getApprovedCaipNetworksData: () => Promise<{
+    approvedCaipNetworkIds: NetworkControllerState['approvedCaipNetworkIds']
+    supportsAllNetworks: NetworkControllerState['supportsAllNetworks']
+  }>
+}
+
+export interface NetworkControllerState {
+  _client?: NetworkControllerClient
+  supportsAllNetworks: boolean
+  requestedCaipNetworks?: CaipNetwork[]
+  approvedCaipNetworkIds?: CaipNetworkId[]
+  allowUnsupportedCaipNetwork?: boolean
+  smartAccountEnabledNetworks?: number[]
 }
 
 export type ChainAdapter = {
