+++ packages/scaffold-ui/src/views/w3m-email-verify-otp-view/index.ts
@@ -5,7 +5,6 @@ import {
   EventsController,
   ConnectionController,
   ModalController,
-  NetworkController,
   RouterController,
   AccountController,
   ChainController,
@@ -35,7 +34,7 @@ export class W3mEmailVerifyOtpView extends W3mEmailOtpWidget {
   override onOtpSubmit: OnOtpSubmitFn = async otp => {
     try {
       if (this.authConnector) {
-        const smartAccountEnabled = NetworkController.checkIfSmartAccountEnabled()
+        const smartAccountEnabled = ChainController.checkIfSmartAccountEnabled()
         await this.authConnector.provider.connectOtp({ otp })
         EventsController.sendEvent({ type: 'track', event: 'EMAIL_VERIFICATION_CODE_PASS' })
