+++ packages/wallet/src/W3mFrameProvider.ts
@@ -91,7 +91,6 @@ export class W3mFrameProvider {
 
   public async getChainId() {
     try {
-      console.log('>> W3mFrameProvider getChainId')
       const response = await this.appEvent<'GetChainId'>({
         type: W3mFrameConstants.APP_GET_CHAIN_ID
       } as W3mFrameTypes.AppEvent)
